+++ backend/src/main/kotlin/com/app/backend/global/config/RedisConfig.kt
@@ -8,6 +8,9 @@ import com.fasterxml.jackson.datatype.jdk8.Jdk8Module
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule
 import com.fasterxml.jackson.module.kotlin.KotlinModule
 import com.fasterxml.jackson.module.paramnames.ParameterNamesModule
+import org.redisson.Redisson
+import org.redisson.api.RedissonClient
+import org.redisson.config.Config
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
@@ -22,8 +25,8 @@ import org.springframework.data.redis.serializer.StringRedisSerializer
 @Configuration
 @EnableRedisRepositories
 class RedisConfig(
-    @Value("\${spring.data.redis.host}") private val host: String,
-    @Value("\${spring.data.redis.port}") private val port: Int,
+    @Value("\${spring.data.redis.host:localhost}") private val host: String,
+    @Value("\${spring.data.redis.port:6379}") private val port: Int,
     @Value("\${spring.data.redis.password}") private val passwordStr: String
 ) {
     @Bean
@@ -49,4 +52,17 @@ class RedisConfig(
                 )
             })
         }
+
+    @Bean
+    fun redissonClient(): RedissonClient {
+        val config = Config()
+        if (passwordStr.isNotBlank()) config.useSingleServer()
+            .setAddress("redis://$host:$port")
+            .setPassword(passwordStr)
+            .setConnectionPoolSize(64)
+        else config.useSingleServer()
+            .setAddress("redis://$host:$port")
+            .setConnectionPoolSize(64)
+        return Redisson.create(config)
+    }
 }
